revamp campaign page with tabbed sections
